backend/requirements.txt
fastapi
google-genai
pydantic
python-dotenv
uvicorn
